Rename stub_query_processor -> query_processor_stub
