+++ irohad/torii/processor/query_processor_stub.hpp
